# packages/wunderland/apps/frame.dev/public/openapi.yaml
openapi: 3.1.0
info:
  title: Quarry Codex API
  description: |
    REST API for Quarry Codex knowledge management system.

    ## Features
    - **Blocks API**: Create, read, update, delete content blocks
    - **Generate API**: LLM-powered text and image generation
    - **Analysis API**: NLP analysis and summarization
    - **GitHub API**: Repository integration and sync

    ## Authentication
    Most endpoints require authentication via API key or session cookie.
    Set `Authorization: Bearer <token>` header for authenticated requests.

    ## Client-Side AI
    Many features run entirely client-side using:
    - **BERT embeddings** via Transformers.js
    - **TextRank summarization** in Web Workers
    - **TF-IDF keyword extraction**

    No data leaves your device for these operations.
  version: 1.0.0
  contact:
    name: Quarry Support
    url: https://frame.dev/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /api
    description: Local API server
...
paths:
...
  /images/generate:
    post:
      tags: [Images]
      summary: Generate image
      description: Generate images using DALL-E or other providers
      operationId: generateImage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prompt]
              properties:
                prompt:
                  type: string
                  description: Image generation prompt
                size:
                  type: string
                  enum: ['1024x1024', '1792x1024', '1024x1792']
                  default: '1024x1024'
                quality:
                  type: string
                  enum: [standard, hd]
                  default: standard
                style:
                  type: string
                  enum: [natural, vivid]
                  default: natural
                provider:
                  type: string
                  enum: [openai, stability]
                  default: openai
      responses:
        '200':
          description: Generated image
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImageResult'
...
  /jobs/{jobId}:
    get:
      tags: [Jobs]
      summary: Get job status
      description: Get status of a specific job
      operationId: getJob
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
components:
  schemas:
...
    ImageResult:
      type: object
      properties:
        url:
          type: string
          format: uri
        base64:
          type: string
        revisedPrompt:
          type: string
        cost:
          type: number
        provider:
          type: string
        model:
          type: string
...
    Job:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
          minimum: 0
          maximum: 100
        result:
          type: object
        error:
          type: string
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
  - bearerAuth: []
  - apiKey: []